Rename parse_funcs into parse module
